requirements.txt
pillow
tensorflow
keras
numpy
matplotlib
ttkthemes
scipy